condensed main game screen classes, updated fonts
