+++ source/core/assets/sounds/ui/open_close/close_01.ogg
