+++ source/core/assets/sounds/ui/open_close/open_01.ogg
