+++ source/core/src/main/com/csse3200/game/components/gamearea/CurrencyDisplay.java
@@ -2,7 +2,6 @@ package com.csse3200.game.components.gamearea;
 
 import com.badlogic.gdx.audio.Sound;
 import com.badlogic.gdx.graphics.Camera;
-import com.badlogic.gdx.graphics.Color;
 import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.BitmapFont;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
@@ -16,7 +15,6 @@ import com.badlogic.gdx.scenes.scene2d.ui.TextButton;
 import com.badlogic.gdx.scenes.scene2d.utils.Drawable;
 import com.badlogic.gdx.scenes.scene2d.utils.TextureRegionDrawable;
 import com.badlogic.gdx.utils.Align;
-import com.csse3200.game.services.GameTime;
 import com.csse3200.game.services.ServiceLocator;
 import com.csse3200.game.ui.UIComponent;
 import com.badlogic.gdx.scenes.scene2d.actions.Actions;
@@ -30,6 +28,7 @@ public class CurrencyDisplay extends UIComponent {
     private TextButton scrapsTb;
     private TextButton crystalsTb;
     private Sound clickSound;
+    private static final String defaultFont = "glitch_24";
 
     /**
      * Adds actors to stage
@@ -64,7 +63,7 @@ public class CurrencyDisplay extends UIComponent {
     private TextButton createButton(String imageFilePath, int value) {
         Drawable drawable = new TextureRegionDrawable(new TextureRegion(new Texture(imageFilePath)));
         TextButton.TextButtonStyle style = new TextButton.TextButtonStyle(
-                drawable, drawable, drawable, new BitmapFont());
+                drawable, drawable, drawable, getSkin().getFont(defaultFont));
 
         // create button
         TextButton tb = new TextButton(String.format("%d", value), style);
