+++ source/core/src/main/com/csse3200/game/components/maingame/MainGameActions.java
@@ -11,7 +11,7 @@ import org.slf4j.LoggerFactory;
  */
 public class MainGameActions extends Component {
   private static final Logger logger = LoggerFactory.getLogger(MainGameActions.class);
-  private GdxGame game;
+  private final GdxGame game;
 
   public MainGameActions(GdxGame game) {
     this.game = game;
