--- source/core/src/main/com/csse3200/game/components/pausemenu/PauseMenuSettingsButton.java
@@ -1,71 +1,0 @@
-package com.csse3200.game.components.pausemenu;
-
-import com.badlogic.gdx.Game;
-import com.badlogic.gdx.graphics.g2d.SpriteBatch;
-import com.badlogic.gdx.scenes.scene2d.Actor;
-import com.badlogic.gdx.scenes.scene2d.ui.Table;
-import com.badlogic.gdx.scenes.scene2d.ui.TextButton;
-import com.badlogic.gdx.scenes.scene2d.utils.ChangeListener;
-import com.csse3200.game.GdxGame;
-import com.csse3200.game.ui.ButtonFactory;
-import com.csse3200.game.ui.UIComponent;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-/**
- * Displays a button in the pause menu to open the game settings.
- */
-public class PauseMenuSettingsButton extends UIComponent {
-    private static final Logger logger = LoggerFactory.getLogger(PauseMenuSettingsButton.class);
-    private static final float Z_INDEX = 2f;
-    private Table table;
-
-    private final GdxGame game;
-
-    public PauseMenuSettingsButton(GdxGame screenSwitchHandle) {
-        game = screenSwitchHandle;
-    }
-
-    @Override
-    public void create() {
-        super.create();
-        addActors();
-    }
-
-    private void addActors() {
-        table = new Table();
-        table.top().right();
-        table.setFillParent(true);
-        TextButton pauseMenuBtn = ButtonFactory.createButton("Settings");
-
-        // Triggers an event when the button is pressed.
-        pauseMenuBtn.addListener(
-                new ChangeListener() {
-                    @Override
-                    public void changed(ChangeEvent changeEvent, Actor actor) {
-                        logger.debug("Settings button clicked");
-                        game.setScreen(GdxGame.ScreenType.SETTINGS);
-                    }
-                });
-
-        table.add(pauseMenuBtn).padTop(300f).padRight(500f);
-
-//        stage.addActor(table);
-    }
-
-    @Override
-    public void draw(SpriteBatch batch) {
-        // draw is handled by the renderer component
-    }
-
-    @Override
-    public float getZIndex() {
-        return Z_INDEX;
-    }
-
-    @Override
-    public void dispose() {
-        table.clear();
-        super.dispose();
-    }
-}
