+++ source/core/src/main/com/csse3200/game/screens/AssetLoader.java
@@ -176,11 +176,11 @@ public class AssetLoader {
             "sounds/mobBoss/patrickCast.mp3",
             "sounds/mobBoss/patrickThunder.mp3",
             "sounds/mobBoss/patrickHit.mp3",
-            "sounds/ui/Click/NA_SFUI_Vol1_Click_01.ogg",
-            "sounds/ui/Hover/NA_SFUI_Vol1_hover_01.ogg",
-            "sounds/ui/Open_Close/NA_SFUI_Vol1_Close_01.ogg",
-            "sounds/ui/Open_Close/NA_SFUI_Vol1_Open_01.ogg",
-            "sounds/ui/Switch/NA_SFUI_Vol1_switch_01.ogg",
+            "sounds/ui/click/click_01.ogg",
+            "sounds/ui/hover/hover_01.ogg",
+            "sounds/ui/open_close/close_01.ogg",
+            "sounds/ui/open_close/open_01.ogg",
+            "sounds/ui/switch/switch_01.ogg",
             "sounds/background/desert/Elements.ogg",
             "sounds/background/desert/Rocks1.ogg",
             "sounds/background/desert/Rocks2.ogg",
