+++ source/core/src/main/com/csse3200/game/screens/MainGameScreen.java
@@ -17,7 +17,7 @@ import com.csse3200.game.areas.ForestGameArea;
 import com.csse3200.game.components.gamearea.PerformanceDisplay;
 import com.csse3200.game.components.maingame.MainGameActions;
 import com.csse3200.game.components.maingame.MainGameLoseDisplay;
-import com.csse3200.game.components.maingame.MainGamePauseDisplay;
+import com.csse3200.game.components.maingame.MainGameDisplay;
 import com.csse3200.game.entities.Entity;
 import com.csse3200.game.entities.EntityService;
 import com.csse3200.game.entities.factories.RenderFactory;
@@ -29,7 +29,6 @@ import com.csse3200.game.rendering.Renderer;
 import com.csse3200.game.services.*;
 import com.csse3200.game.ui.terminal.Terminal;
 import com.csse3200.game.ui.terminal.TerminalDisplay;
-import com.csse3200.game.components.maingame.MainGameExitDisplay;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -56,11 +55,11 @@ public class MainGameScreen extends ScreenAdapter {
           "sounds/background/desert/desert_bgm.ogg"
   };
   private static final String[] uiSounds = {
-          "sounds/ui/Click/NA_SFUI_Vol1_Click_01.ogg",
-          "sounds/ui/Hover/NA_SFUI_Vol1_hover_01.ogg",
-          "sounds/ui/Open_Close/NA_SFUI_Vol1_Close_01.ogg",
-          "sounds/ui/Open_Close/NA_SFUI_Vol1_Open_01.ogg",
-          "sounds/ui/Switch/NA_SFUI_Vol1_switch_01.ogg"
+          "sounds/ui/click/click_01.ogg",
+          "sounds/ui/hover/hover_01.ogg",
+          "sounds/ui/open_close/close_01.ogg",
+          "sounds/ui/open_close/open_01.ogg",
+          "sounds/ui/switch/switch_01.ogg"
   };
   private static final String[] desertSounds = {
           "sounds/background/desert/Elements.ogg",
@@ -305,10 +304,9 @@ public class MainGameScreen extends ScreenAdapter {
         .addComponent(new PerformanceDisplay())
             .addComponent(new MainGameActions(this.game))
             .addComponent(ServiceLocator.getWaveService().getDisplay())
-            .addComponent(new MainGameExitDisplay())
             .addComponent(new MainGameLoseDisplay())
             //.addComponent(new MainGameWinDisplay()) <- needs to be uncommented when team 3 have implemented the ui
-            .addComponent(new MainGamePauseDisplay(this.game))
+            .addComponent(new MainGameDisplay(this.game))
             .addComponent(new Terminal())
             .addComponent(inputComponent)
             .addComponent(new TerminalDisplay());
